Create uncompleted_player.cpp
ountless marathoners participated in the marathon. All but one athlete completed the marathon.

When you are given an arrangement that contains the names of the athletes who participated in the marathon and an arrangement that contains the names of the finished athletes, write a solution function to return the names of the athletes who did not finish.

Limitations
The number of athletes participating in the marathon is more than one and less than 100,000.
The length of the complete is 1 less than the length of the partition.
Participants' names consist of one or more lowercase alphabetic characters.
Participants may have the same name.