Create List_of_telephone_numbers
https://programmers.co.kr/learn/courses/30/lessons/42577

I want to check if one of the phone numbers in the phone book is a prefix for another number.
If the phone number is as follows, the rescue phone number is the prefix of Young-seok's phone number.

Rescue: 119
Park Joon-young: 97 674 223
Ji Young-seok: 119552 4421
When an array phone_book containing the phone numbers listed in the phone book is given as a parameter for the solution function, please write the solution function to return false otherwise true.

Limitations
The length of the phone_book is from 1 to 1,000,000.
Each phone number is 1 or less long.
I/O Example
phone_book return
[119, 97674223, 1195524421] false
[123,456,789] True
[12,123,1235,567,88] false
I/O Example Description
I/O Example #1
This is the example described earlier.

I/O Example #2
The answer is true because one number is often the prefix of another number.

I/O Example #3
The first phone number, "12", is the prefix of the second phone number "123". Therefore, the answer is false.