Create Best_Album
Problem description
We are planning to release our best album by collecting the two most played songs by genre on streaming sites. Songs are separated by unique numbers, and the criteria for containing songs are as follows:

We first include the genre in which many of our songs have been played.
We first include songs that have been played a lot in the genre.
Among songs with the same number of plays within the genre, songs with lower unique numbers are included first.
When a string array genres representing the genre of the song and integer array plays showing the number of plays per song are given, complete the solution function to turn the unique numbers of the songs to be included in the best album in order.

Limitations
Genres[i] is a genre of songs with a unique number i.
Plays[i] is the number of times a song with the unique number i has been played.
The genres and plays are the same length, which is from 1 to 10,000.
There are less than 100 genres.
If there is only one song in the genre, select only one song.
Every genre has a different number of plays.
I/O Example
genres plays return
[classic, pop, classic, classic, pop] [500, 600, 150, 800, 2500] [4, 1, 3, 0]
I/O Example Description
The classic genre has been played 1,450 times, and the classic song is as follows.

Unique number 3: 800 playback
Unique number 0: 500 plays
Unique number 2: Play 150 times
The pop genre has been played 3,100 times, and the pop song is as follows.

Unique number 4: 2,500 playback
Unique number 1: 600 playback
Therefore, the pop genre's [4, 1] song first, the classic genre's [3, 0] song, and then the next song.