Create K_number
You want to get the number in the kth when you cut and sort from the i-th digit to the j-th digit of the array array.

For example, if array is [1, 5, 2, 6, 3, 7, 4], i = 2, j = 5, k = 3,

[5, 2, 6, 3] is cut from second to fifth of array.
Arrangement from 1 is [2, 3, 5, 6].
The third number of arrays from 2 is 5.
When the two-dimensional array commands with array array, [i, j, k] are given as parameters, write a solution function to return the results from the operation described earlier for all elements of command.

Limitations
Array is 1 to 100 long.
Each element of array is 1 to 100 or less.
Commands are 1 to 50 long.
Each element of the commands is 3 long.
I/O Example
array commands return
[1, 5, 2, 6, 3, 7, 4] [[2, 5, 3], [4, 4, 1], [1, 7, 3]] [5, 6, 3]
I/O Example Description
Cut [1, 5, 2, 6, 3, 7, 4] from 2nd to 5th, and then arrange it. The third number of [2, 3, 5, 6] is 5.
Cut [1, 5, 2, 6, 3, 7, 4] from fourth to fourth, and then align. The first number of [6] is 6.
Cut [1, 5, 2, 6, 3, 7, 4] from 1st to 7th. The third number of [1, 2, 3, 4, 5, 6, 7] is 3.