Create H_Index
Problem description
H-Index is an indicator of the productivity and influence of scientists. You are trying to obtain a value of h that represents the H-Index of a scientist. According to Wikipedia1, H-Index is:

Of the n papers published by a scientist, if more than h has been cited and the rest has been cited less than h, the maximum value of h is the scientist's H-Index.

When array cits containing the number of quotes in a paper published by a scientist are given as parameters, please write a solution function to return this scientist's H-Index.

Limitations
The number of papers published by scientists is more than one and less than 1,000.
The number of quotes per paper is 0 or less than 10,000 times.
I/O Example
cites return
[3, 0, 6, 1, 5] 3
I/O Example Description
The number of papers published by this scientist is five, and three of them have been cited more than three times. And because the other two papers were quoted less than three times, the scientist's H-Index is three.