Create Function_Development
Problem description
The Promptus team is working on a functional improvement. Each feature can be reflected in the service at 100% progress.

In addition, because each function has a different development speed, the function behind it can be developed before the function in front of it, which is then distributed together when the function in front of it is distributed.

Complete the solution function to reset how many functions are deployed for each deployment when an integer array progresses in the order in which the tasks should be deployed first and an integer array speed in which each task's development speed is given.

Limitations
The number of operations (progresses, length of speeds array) is less than or equal to 100.
The work progress is natural water less than 100.
The working speed is 100 or less natural water.
It is assumed that the deployment can be done only once a day, and is done at the end of the day. For example, if a job with a 95% progress rate is 4% per day, the deployment will take two days.
I/O Example
pressure speeds return
[93,30,55] [1,30,5] [2,1]
I/O Example Description
The first feature is 93% complete and can be worked on 1% per day, so it can be deployed after 7 days of work.
The second feature is 30% complete, and you can work 30% per day, so you can deploy it after working for 3 days. However, because the previous first feature is not yet complete, it will be deployed on the 7th day of deployment.
The third feature is 55% complete and can be worked on 5% per day, so it can be deployed after 9 days of work.

Therefore, two functions are distributed on day 7 and one on day 9.