Create A_bridge_truck
a bridge truck
Problem description
Several trucks are about to cross the first lane bridge across the river in order. Find out how many seconds it takes for all trucks to cross the bridge. The truck moves by 1 second, the length of the bridge_length, and the legs stand to weight weight.
※ If the truck is not fully loaded on the bridge, the weight of the truck is not considered.

For example, there is a bridge that is 2 long and weighs 10kg. For trucks weighing [7, 4, 5, 6]kg to cross the bridge in the shortest time in order:

a truck waiting to cross a bridge past its elapsed time.
0 [] [] [7,4,5,6]
1~2 [] [7] [4,5,6]
3 [7] [4] [5,6]
4 [7] [4,5] [6]
5 [7,4] [5] [6]
6~7 [7,4,5] [6] []
8 [7,4,5,6] [] []
Therefore, it takes at least 8 seconds for all trucks to pass the bridge.

The parameters of the solution function are leg length bridge_length, weight weight that the leg can withstand, and weight truck_weights per truck. At this point, complete the solution function so that all trucks can return at least a few seconds to cross the bridge.

limiting condition
Bridge_length is 1 or more and less than 10,000.
Weight is from 1 to 10,000.
The length of the truck_weights is 1 or more and less than 10,000.
All trucks weigh more than 1 weight or less.
I/O Example
bridge_length weight truck_weights return
2 10 [7,4,5,6] 8
100 100 [10] 101
100 100 [10,10,10,10,10,10,10,10,10,10] 110