Create printer
Typical printers print in the order in which print requests are received. This is why important documents can be printed later. To compensate for this problem, we have developed a printer that prints high-priority documents first. This newly developed printer performs printing in the following ways:

1. Remove the document (J) at the front of the print waitlist from the waitlist.
2. If any of the remaining print wait lists have more important documents than J, place J at the end of the wait list.
3. Otherwise, print J.
For example, four documents (A, B, C, D) will be on the print waiting list in order, followed by 2 1 3 2 instant noodles C D A B.

I'd like to know the number of times the document I requested to print is printed. In the example above, C is printed first and A is printed third.

Please write a solution function to reset the number of documents I requested to print when the parameters of the current list of documents are given the order of importance and the location that tells me where the documents I requested are currently on the waiting list.

Limitations
Currently, there are more than 1 document or less than 100 documents in the waiting list.
The importance of print jobs is expressed in 1 to 9, which means that larger numbers are more important.
Location has a value greater than 0 (number of jobs in the current waitlist - 1) and is expressed as 0 at the front of the wait list and 1 at the second.
I/O Example
privileges location return
[2, 1, 3, 2] 2 1
[1, 1, 9, 1, 1, 1] 0 5
I/O Example Description
Example #1

This is an example of a problem.

Example #2

The six documents (A, B, C, D, E, F) are on the print waiting list and are important to 1 9 1 1 1 1 1 so print C D E F A B.