Update network
image : https://programmers.co.kr/learn/courses/30/lessons/43162#
Problem description
A network is a connected form that allows you to exchange information between computers. For example, when computer A and computer B are directly connected, computer B and computer C are directly connected, computer A and computer C can also be indirectly connected to exchange information. Therefore, computers A, B, and C are all on the same network.

When the number of computers n, two-dimensional array computers with information about connections are given as parameters, write a solution function to return the number of networks.

Limitations
The number n of computers is a natural number from 1 to 200.
Each computer is represented by an integer from 0 to n-1.
If computer i and computer j are connected, describe computers[i][j] as 1.
Computer[i][i] is always 1.
I/O Example
n computers return
3 [[1, 1, 0], [1, 1, 0], [0, 0, 1]] 2
3 [[1, 1, 0], [1, 1, 1], [0, 1, 1]] 1