Create disk_controller
Problem description
A hard disk can only perform one operation at a time. There are several ways to implement a disk controller. The most common method is to process requests in the order in which they are received.

image : https://programmers.co.kr/learn/courses/30/lessons/42627#

for example

- 3 ms required for A operation at 0 ms time
- Requests work B that takes 9 ms at 1 ms.
- 6 ms required for C operation at 2 ms time
A request has been received, such as . This is illustrated below.
Screen Shot 2018-09-13 at 6.34.58 PM.png

Because only one request can be made at a time, processing each action in the order in which it is requested will result in:
Screen Shot 2018-09-13 at 6.38.52 PM.png

- A: Work completed at 3 ms (request to end: 3 ms)
- B: Wait from 1 ms, start work at 3 ms, complete work at 12 ms (request to end: 11 ms)
- C: Wait from 2 ms, start work at 12 ms, complete work at 18 ms (request to end: 16 ms)
The average time taken from the request to the end of each task is 10 ms (3 + 11 + 16) / 3.

However, if A → C → B is processed in order,
Screen Shot 2018-09-13 at 6.41.42 PM.png

- A: Work completed at 3 ms (request to end: 3 ms)
- C: Wait from 2 ms, start work at 3 ms, complete work at 9 ms (request to end: 7 ms)
- B: Wait from 1 ms, start work at 9 ms, complete work at 18 ms (request to end: 17 ms)
This order of A C C B B results in an average of 9 ms (3 + 7 + 17) of the time taken from request to end of each operation.

For each task, when a two-dimensional array job is given as a parameter, write a solution function to return the average of the time taken from request to end of the task in the most reduced way. (However, throw away any number below the decimal point)

Limitations
Jobs are 1 to 500 long.
Each row of jobs is [when the job is requested, time taken].
For each job, the time the task is requested is from 0 to 1000.
For each job, the time required is from 1 to 1000.
If the hard disk is not performing an operation, it first processes the requested operation.
I/O Example
jobs return
[[0, 3], [1, 9], [2, 6]] 9
I/O Example Description
This is the example given to a problem.

You receive a 3 ms operation request at 0 ms time.
You receive a 9 ms request at 1 ms time.
You will receive a 6 ms request at 2 ms time.