Create double_priority_Q
double-priority Q
Problem description
Double-priority queue refers to a data structure that can perform the next operation.

Command Receive Top (High)
I insert a given number in the numeric queue.
D 1 Deletes the maximum value from the queue.
D -1 Delete the maximum value from the queue.
When the operations to be double-priority queue are given as parameters, if the queue is empty after processing all operations [0,0] If it is not empty, implement the solution function to return [maximum, maximum value].

Limitations
operations are an array of strings that is 1 or more long and less than 1,000,000.
The elements of the operations represent the operations that the queue will perform.
Element is given in the form of "command data".- If there is more than one maximum/maximum value in the operation to delete the maximum/maximum value, delete only one.
If an empty queue is given an operation to delete the data, the operation is ignored.
I/O Example
operations return
[I 16,D 1] [0,0]
[I 7,I 5,I -5,D -1] [7,5]
I/O Example Description
Insert 16 and delete the maximum value. Returns [0,0] because it is empty.
Insert 7,5,-5 and delete the maximum value. Returns the maximum value of 7, and the minimum value of 5.

// ps : It says that the problem is related to the priority queue, but because it is sufficient to solve it using vectors, we solved the problem using vectors.