Update Life_boat
They try to rescue people who are trapped on a deserted island by using a lifeboat. The lifeboat is small, so you can only ride two people at a time, and there is a weight limit.

For example, if people weigh [70kg, 50kg, 80kg, 50kg] and the weight limit of the lifeboat is 100kg, the second and fourth person can ride together, but the sum of the weights of the first and third person is 150kg, so they cannot ride together beyond the weight limit of the lifeboat.

Try to save everyone by using as few lifeboats as possible.

When the weight limit limit of the people and the weight limit of the lifeboat are given as parameters, write a solution function to reset the maximum number of lifeboats required to rescue everyone.

Limitations
There are more than 1 person or less than 50,000 people trapped on a deserted island.
Each person weighs more than 40kg and less than 240kg.
The weight limit of lifeboats is 40kg to 240kg.
The weight limit of lifeboats is always given above the maximum weight of a person, so there is no case that people cannot be rescued.
I/O Example
people limit return
[70, 50, 80, 50] 100 3
[70, 80, 50] 100 3

ps./ I don't know why this is a Greedy Approach problem.
However, I have decided to put the folder in question here because the classification suggested by the examiner in question is divided into Greedy Approach.
I think this problem is solved using Sort rather than Greedy Approach.