Update Connecting Isles
image -> https://programmers.co.kr/learn/courses/30/lessons/42861#
Complete the solution to return the minimum cost of building bridges between n islands and making all islands accessible to each other at a minimum cost.

Even if you cross the bridge several times, you can get there if you can reach it. For example, if there is a bridge between Island A and Island B, and if there is a bridge between Island B and Island C, Island A and Island C can pass each other.

Limitations

The number of isles n is 1 to 100 or less.
The length of the cost is (n-1) * n) / less than 2.
For any i, cos[i][0] and coss[i] [1] contain the numbers of the two islands to which the bridge is connected, and cost of building the bridge to connect the two islands to the coss [i] [2].
The same connection is not given twice. Also, if the order changes, view it as the same connection. That is, when there is a cost between 0 and 1, there is no cost of 1 and 0.
There is no cost to build bridges between all islands. In this case, construction between the two islands is considered impossible.
No unreachable islands are given.
I/O Example

n costs return
4 [[0,1,1],[0,2,2],[1,2,5],[1,3,1],[2,3,8]] 4
I/O Example Description

Illustration of the costs shows that connecting to the green path is the way to get everyone through at the lowest cost.