Update Immigration
N people are waiting in line for immigration. Each examiner at each immigration checkpoint takes a different amount of time.

At first, all the screening tables are empty. Only one judge can judge at the same time. The person standing in front of you can go to an empty examiner's desk and be judged. However, if there is an examiner who finishes earlier, you can wait and go there to get the screening.

I want to minimize the time it takes for everyone to be screened.

Number of people waiting for immigration, n When each examiner is given an array time as a parameter, write a solution function to reset the maximum value of time it takes everyone to be screened.

Limitations
The number of people waiting for immigration is 1 to 1,000,000,000.
Each examiner takes more than one minute to judge one person, not more than 1,000,000,000 minutes.
The number of examiners is 1 to 100,000.
I/O Example
n times return
6 [7, 10] 28
I/O Example Description
The first two go straight to the screening.

By seven minutes, the first panel will be empty and the third person will be judged.

By 10 minutes, the second panel will be empty and the fourth person will be judged.

By 14 minutes, the first panel will be empty and the fifth person will be judged.

At 20 minutes, the second panel is empty, but if the sixth person waits another minute without being screened there, and then the first panel judges, everyone's screening will be completed at 28 minutes.