Create stepping_stone
Problem description
There is an arrival point as far as the distance from the departure point. And in between, there are rocks. You are about to remove some of the rocks.
For example, if you remove two rocks when the arrival point is 25 and the rock is located [2, 14, 11, 21, 17], the distance between the starting point, the destination, and the rocks is as follows.

Position of Rock Removed The maximum distance between each rock
[21, 17] [2, 9, 3, 11] 2
[2, 21] [11, 3, 3, 8] 3
[2, 11] [14, 3, 4, 4] 3
[11, 21] [2, 12, 3, 8] 2
[2, 14] [11, 6, 4, 4] 4
The largest value of the distance obtained above is 4.

When the distance from the starting point to the arrival point, the array rocks containing the locations of the rocks, and the number n of rocks to remove are given as parameters, remove n rocks and create a solution function to return the largest of the maximum values of the distance between each point.

Limitations
The distance to the destination is from 1 to 1,000,000,000,000 won.
There are more than one rock and less than 50,000 rocks.
n is more than 1 and less than the number of rocks.
I/O Example
distance rocks n return
25 [2, 14, 11, 21, 17] 2 4

1. The maximum and maximum values of the values to be obtained should be determined according to the problem of dichotomy. The value to be sought is to obtain the highest value of the distance between the remaining rocks after removing as many as n rocks. In other words, it is a matter of distance between rocks.

2. The maximum value of distance between rocks is 1. The maximum value is set to the given variable, distance.

3. If the middle value is obtained, the rock shall be removed if the distance of the rock is less than the mid value obtained by touring the arrangement containing the location of the rock, and if it is equal or greater than the mid value obtained, the rock shall not be removed. Count the number of rocks removed and save the index of the rocks if not removed.
(The reason is to calculate the distance between the rock that has not been removed and the rock selected in the next (a).

4. If the last number of rocks removed is greater than n, correct the right position because the distance between the rocks (mid) specified by us is too large and must be reduced. Conversely, if the number of rocks removed is small, the location of the left should be corrected because the distance between the rocks is too small.

The points to be noted in No. 4 are as follows.
After changing the condition statement of 30 line to count <= and 32 line to else, the correct answer is not obtained when the part of 34 line is inserted inside the if statement of 30 line. If the 30 line renews the anwer while proceeding with whim, the maximum value will be found while the distance between rocks is increased by removing many rocks.
Therefore, we coded lines 30 to 34 to implement the method of finding the maximum value that can exist when less than n rocks are removed.