explain
image -> https://programmers.co.kr/learn/courses/30/lessons/49189

There is a graph with n nodes. Each node has a number from 1 to n. You are trying to obtain the number of nodes farthest from node 1. The furthest node means the nodes with the largest number of intersections when moved to the shortest path.

When the parameter is given a 2-dimensional array vertex with information about the number of nodes n, the main line, write a solution function to return how many nodes are farthest from node 1.

Limitations
The number of nodes n is 2 or more and less than 20,000.
The main lines are bi-directional, with a total of 1 to 50,000 main lines.
Vertex array Each row [a, b] means that there is a line between node a and node b.
I/O Example
n vertex return
6 [[3, 6], [4, 3], [3, 2], [1, 3], [1, 2], [2, 4], [5, 2]] 3
I/O Example Description
The graph in the example is shown below, and the nodes farthest from node 1 are nodes 4, 5, and 6.