Update The_furthest_node_bfs
The reason why the code before the modification took a long time seems to be because of the parameters passed to the bfs function.
In the existing bfs function, there was a two-dimensional vector parameter called map, which was treated as a global variable and accelerated.
But treating array or quee as a global variable actually reduced the speed. I think we should learn about this by studying.