Create ranking
N boxers participated in the boxing competition and each received numbers from number 1 to n. The boxing match is carried out in a one-on-one manner, and if A is better than B, A always beats B. The referee wants to rank the players with the results of the given competition. However, the results of some matches have been lost and cannot be ranked accurately.

Write a solution function so that the number of competitors n, two-dimensional array results with the results of the competition, can be accurately ranked when given as parameters.

Limitations
The number of athletes is between 1 and 100 players.
The result is more than 1 and less than 4,500.
Results Array Each row [A, B] means that A beat B.
There is no contradiction in all competitions.
I/O Example
n results return
5 [[4, 3], [4, 2], [3, 2], [1, 2], [2, 5]] 2
I/O Example Description
Player 2 is in fourth place because he lost to [1, 3 and 4] and won to player 5.
Player No. 5 is No. 5 because he lost to No. 4 player No. 2.